workshop-1-production/Code/copilot_dem/requirements.txt
pydantic-ai 
openai-agents 
nest-asyncio 
streamlit